+++ test/raylib/test_runner.clj
@@ -55,6 +55,7 @@
                      raylib.scenes.splines-test
                      raylib.scenes.rounded-test
                      raylib.scenes.vecangle-test
+                     raylib.scenes.bars-test
                      raylib.easings-test
                      raylib.scenes.clock-test
                      raylib.scenes.easings-test